requirements.txt
fastapi
uvicorn
transformers
torch
python-dotenv
pycryptodome
